Refactor plugin api
The plugin api has been renamed from "plugins" to "plugin" to avoid an
issue in the core of Sublime Text.

Sublime Text autoloads all modules in the root of a package and
instantiates all Sublime Text commands that it finds. It stores the
collection of commands it finds in a module attribute: "plugins". This
causes an issue for any module definitions or imports named "plugins".

One issue is an error when Sublime Text tries to unload a module that
imports a module named "plugins" e.g. Sublime Text loads the module,
finds any commands, instantiates them and assigns them to an attribute
named "plugins" on the module. If the module then imports a "plugins"
module, it overwrites the collection of commands. When Sublime tries to
unload the module it will error because the collection of commands has
been overwritten.