Handle upgrades gracefully
Redesigned to handle errors gracefully.

When upgrading the plugin, changes to the plugin structure can cause
import errors. In these cases we want to notify the user about needing
to restart Sublime Text to finish the upgrade.

In the case any errors we also don't want to leave the normal
functioning of the editor unusable. We can't access the sublime api
until the plugin_loaded() hook is called, so we need to catch exceptions
and run cleanup operations when the plugin_loaded() hook.