Fix default mode assertions can sometimes fail
Apparently it can fail on the CI servers aswell as local development
machines, though at the time of writing I wasn't able to reproduce the
issue on the CI servers.

The issue:

The default mode is UNKNOWN. The event listener on_activated event
initialises views and part of that initialisation is to set the
default mode to NORMAL.

An issue arises when the Sublime Text window doesn't have focus e.g.
start a test run, then quickly move the focus to another application
while the tests are run.

When ST doesn't have focus it doesn't fire the on_activated event. I
assume it's expected ST behaviour e.g. maybe because the when you move
focus away from ST is fires the on_deactivate event on the active
view.

When the on_activated event is NOT fired for the test view then the
mode will stay in an UNKNOWN state, but when it's fired it is set to
a NORMAL state.

Fixes #129