Internal: Fix VISUAL mode test fixtures
This is required, because the cursor in VISUAL mode is a block cursor.
Without this setting some tests will pass when the window running the
tests has focus, and fail when it doesn't have focus. This happens
because ST doesn't fire events for views (test views) when the window
loses focus: the on_activated() event fixes VISUAL mode selections that
don't have a correct caret state.

Relates to #314