Fix #47 Inserts text behind cursor
NeoVintageous resets to normal mode when a view is activated. This is
done via an on_activated() event.

When the Command Palette is invoked it triggers an on_deactivated()
event on the active view, then on_activated() event on the Command
Palette view.  When Enter or Escape is then pressed from the Command
Palette it triggers an on_deactivated() event on the Command Palette
view, then on_activated() event on the previously active view.

So... when a command is run via the Command Palette, the view is reset
to normal mode before the command has time to run.

This is only an issue for commands run via the Command Palette when the
cursor is at the end of a line, because in Vim when the cursor (a block
cursor) is at the eol, it is actually on the character at the eol:

    ab|c|

The cursor position is actually before the last character. To enter
insert mode after the last character press a.

If in insert mode at the eol:

    abc|

Running a command from Command Palette will reset to normal mode before
the command is run:

    ab|c|

The command then inserts in the wrong position i.e. before the last
character.

The workaround fix is to wrap the Command Palette command in a
NeoVintageous command and ensure that the cursor is not reset to normal
mode.

In previous versions if Vintageous a timer was used to indicate that the
previous view was not the Command Palette:

https://github.com/NeoVintageous/NeoVintageous/commit/f5393ed6a4207a380995b295ed6c4a45d857b660

At the time I removed it I didn't know what the timer was for. I think
the current workaround fix is a better solution until:

In Vim, the cursor is generally not reset to normal mode. When you enter
insert mode and move to another view, you stay in insert mode. When you
enter visual mode & move to another view the visual mode is cleared and
changes to normal mode and then the cursor moves to the other view and
stays in normal mode. Making NeoVintageous work like this will require
significant resign of how the mode state is handled.