Fix and refactor edge-case mapping issues
seq_to_mapping() should return mapping on only when the sequence is
fully matched. Currently it was returning the mapping for the first key
if the sequence didn't match, which is redundant because
is_incomplete() function called before the resolve command, which means
the code that expands the first mapping can never be reached.

Works towards #433