Fix focus causes visual mode to enter normal mode issues
Mapping super keys to visual mode doesn't work because Sublime loses
focus when super keys are press (at least on Ubuntu).

The reason this happens is because when the focus goes back to Sublime
it fires the activated view event, NeoVintageous then clears the visual
selection, which then means the mapping resolver runs in Normal mode and
because the mapping is in visual mode it isn't found.

Possibly relates to: #277 #357 #374

Fixes #547
Fixes #548