Implement search modifiers #641
Some characters in the pattern are taken literally. They match with the
same character in the text. When preceded with a backslash however, these
characters get a special meaning. See :help magic for more details.

Patterns can be prefixed by a "mode" that overrides the 'magic' option:

\m    'magic' on for the following chars in the pattern.
\M    'magic' off for the following chars in the pattern.
\v    the following chars in the pattern are "very magic".
\V    the following chars int the pattern are "very nomagic".

Currently both \m and \v operate the same (magic) and both \M and \V
operate the same (nomagic).

Both magic and nomagic modes in ST are not identical to Vim. It's not
practical to implement the exact regex engine from Vim. Nonetheless,
additional enhancements can be implemented on a case by case basis.

magic mode is essentially STs regex mode with some tweaks to match some
of the behaviours from Vim. nomagic mode is essentially a literal
search.

Close #641