Fix #570 Repeated inner object selctions
For example: vitit

The normal method is to select a <tag> until the matching </tag>. For
"at" the tags are included, for "it" they are excluded. But when "it" is
repeated the tags will be included (otherwise nothing would change).
Also, "it" used on a tag block with no contents will select the leading
tag.

Fixes #570
Re #569