Refactor: remove irreversible command
The IrreversibleTextCommand doesn't actually do anything. The documentation says
that commands derived from it will will bypass the undo:

> Base class. The undo stack will ignore commands derived from this class.
> This is useful to prevent global state management commands from shadowing
> commands performing edits to the buffer, which are the important ones to
> keep in the undo history.
>

That was once the case in a much earlier version, but it was later modified to
fix issues removing the edit token:

> We discard the edit_token because we don't want an
> IrreversibleTextCommand to be added to the undo stack, but Sublime
> Text seems to still require us to begin..end the token. If we removed
> those calls, the caret would blink while motion keys were pressed,
> because --apparently-- we'd have an unclosed edit object around.

See: 41c36707b34ac94277cae481ae4e0bbab5a1bb9c

The end result is that the IrreversibleTextCommand is no different than the
parent TextCommand.