Refactor: state runnable !
Note that the call in add_step() was removed because it wasn't an actual
call but was a function reference which always executed as true. This
has always been the case since 1a59e1c.

The original intent was to only add steps from runnable commands, but
"fixing" that now actually breaks a few things. Macros need reviewing.